refinement: Fix AtsBranchConfigurationTest to check branch existance before purge

+++ plugins/org.eclipse.osee.ats.test/src/org/eclipse/osee/ats/config/AtsBranchConfigurationTest.java
@@ -402,7 +402,9 @@ public class AtsBranchConfigurationTest {
                TestUtil.sleep(2000);
             }
          }
-         BranchManager.purgeBranch(branch);
+         if (BranchManager.branchExists(branch)) {
+            BranchManager.purgeBranch(branch);
+         }
          TestUtil.sleep(2000);
 
       } catch (BranchDoesNotExist ex) {
